Add saving account and change int to float

+++ src/main/java/rocks/zipcode/atm/bank/Account.java
@@ -19,12 +19,12 @@ public abstract class Account {
     }
 
     //takes in a float for parameter
-    public void deposit(int amount) {
+    public void deposit(float amount) {
         updateBalance(getBalance() + amount);
     }
 
     // taking in a float for parameter
-    public boolean withdraw(int amount) {
+    public boolean withdraw(float amount) {
         if (canWithdraw(amount)) {
             updateBalance(getBalance() - amount);
             return true;
@@ -34,16 +34,16 @@ public abstract class Account {
     }
 
     //float again
-    protected boolean canWithdraw(int amount) {
+    protected boolean canWithdraw(float amount) {
         return getBalance() >= amount;
     }
 
-    public int getBalance() {
+    public float getBalance() {
         return accountData.getBalance();
     }
 
     //float
-    private void updateBalance(int newBalance) {
+    private void updateBalance(float newBalance) {
         accountData = new AccountData(accountData.getId(), accountData.getName(), accountData.getEmail(),
                 newBalance);
     }
