+++ src/main/java/rocks/zipcode/atm/bank/AccountData.java
@@ -8,9 +8,9 @@ public final class AccountData {
     private final int id;
     private final String name;
     private final String email;
-    private final int balance;
+    private final float balance;
 
-    AccountData(int id, String name, String email, int balance) {
+    AccountData(int id, String name, String email, float balance) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -29,7 +29,7 @@ public final class AccountData {
         return email;
     }
 
-    public int getBalance() {
+    public float getBalance() {
         return balance;
     }
